copy: _copy_instance: Try to filter class variables.
If field value stored in a class is-matches value stored in an instance,
consider that it's class variable and don't copy it.

This check is heuristic, because instance var may override class var, even
if both have the same value (at particular point of time). We don't copy the
field in this case, and if class var will change later, the instance will
behave differently.